% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_config.R
\name{eval_config}
\alias{eval_config}
\title{Run an algorithm}
\usage{
eval_config(
  learner,
  task_id,
  configuration,
  trainsize,
  logfile = NULL,
  seed = NULL
)
}
\arguments{
\item{learner}{\code{character}\cr
Learner name e.g. "classif.glmnet".}

\item{task_id}{\code{character}\cr
OpenML data ID.}

\item{configuration}{\code{list}\cr
Named list of hyperparameters.}

\item{trainsize}{\code{numeric}\cr
Fraction of training set size.}

\item{logfile}{\code{character}\cr
Logfile to write to.}

\item{seed}{\code{character}\cr
Seed to set. Defaults to \code{NULL} (no custom seed, using original one).}
}
\description{
Run an algorithm.
}
\examples{
learner = "classif.glmnet"
task_id = "40981"
configuration = list(classif.glmnet.alpha = 0.1, classif.glmnet.s = 0.1)
trainsize = 0.05
eval_config(learner, task_id, configuration, trainsize)
}
