% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_config.R
\name{eval_iaml}
\alias{eval_iaml}
\title{Run an algorithm from iaml}
\usage{
eval_iaml(learner, task_id, configuration, ...)
}
\arguments{
\item{learner}{\code{character}\cr
Learner name e.g. "iaml_glmnet".}

\item{task_id}{\code{character}\cr
OpenML data ID.}

\item{configuration}{\code{list}\cr
Named list of hyperparameters.}

\item{...}{\code{any}\cr
Arguments passed on to \code{eval_config}.}
}
\description{
Run an algorithm from iaml.
}
\examples{
learner = "iaml_glmnet"
task_id = "40981"
configuration = list(alpha = 0.1, s = 0.1, trainsize = 0.05)
eval_iaml(learner, task_id, configuration)
}
