% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_config.R
\name{eval_yahpo}
\alias{eval_yahpo}
\title{Run a YAHPO Gym config}
\usage{
eval_yahpo(scenario, configuration, ...)
}
\arguments{
\item{scenario}{\code{character}\cr
Scenario (e.g. \code{iaml_super}) to evaluate.}

\item{configuration}{\code{list}\cr
Named list of hyperparameters including task_id and trainsize.}

\item{...}{\code{any}\cr
Arguments passed on to \code{eval_config}.}
}
\description{
Run a YAHPO Gym config.
}
\examples{
scenario = "iaml_glmnet"
configuration = list(alpha = 0.1, s = 0.1, trainsize = 0.05, task_id = "40981")
eval_yahpo(scenario, configuration)
}
