% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{iaml-package}
\alias{iaml}
\alias{iaml-package}
\title{iaml: Reproduce results from iaml}
\description{
Reproduce results from iaml including seeds.
}
\author{
\strong{Maintainer}: Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})

}
